# backend/src/main/java/com/deutschlingodeck/authentication/dto/LoginRequest.java
package com.deutschlingodeck.authentication.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record LoginRequest(
		@NotBlank @Email String email,
		@NotBlank String password
) {
}