# backend/src/main/java/com/deutschlingodeck/game/dto/AnswerValidationResponse.java
package com.deutschlingodeck.game.dto;

public record AnswerValidationResponse(
		ValidationResult result,
		Boolean correct,
		String givenAnswer,
		String expectedAnswer,
		String message,
		CurrentCardResponse nextCard
) {
}